[refactor] modal component
1. 폴더 구조 분리: base, specific
2. type 폴더 분리
3. PascalCase로 폴더명 변경
