+++ src/shared/ui/modal/base/modal.less
