+++ src/shared/ui/modal/index.ts
@@ -1,7 +1,8 @@
-export { AlertsModal } from './alerts-modal';
-export { EligibilityReviewVerifiedModal } from './eligibility-review-verified-modal';
-export { NameModal } from './name-modal';
-export { PatientConsentModal } from './patient-consent-modal';
-export { ServiceTypeModal } from './service-type-modal';
-export { StartDateModal } from './start-date-modal';
-export { Modal } from './modal';
+export { AlertsModal } from './specific/AlertsModal';
+export { EligibilityReviewVerifiedModal } from './specific/EligibilityReviewVerifiedModal';
+export { NameModal } from '@/shared/ui/modal/specific/NameModal.tsx';
+export { PatientConsentModal } from './specific/PatientConsentModal';
+export { ServiceTypeModal } from './specific/ServiceTypeModal';
+export { StartDateModal } from './specific/StartDateModal';
+export * from '@/shared/ui/modal/base/Modal';
+export * from '@/shared/ui/modal/type'
